chore: Update Type location + update imports
