+++ openwink-app/Components/index.ts
@@ -13,3 +13,4 @@ export * from "./ModuleUpdateModal";
 export * from "./InfoBox";
 export * from "./MiataHeadlights";
 export * from "./ModalBlurBackground";
+export * from "./CustomButtonActionModal";
