+++ openwink-app/Storage/CustomOEMButtonStore.ts
@@ -1,7 +1,6 @@
-import { ButtonBehaviors, Presses } from "../helper/Types";
+import { ButtonBehaviors, CommandOutput, Presses } from "../helper/Types";
 import { buttonBehaviorMap } from "../helper/Constants";
 import Storage from "./Storage";
-import { CommandOutput } from "./CustomCommandStore";
 
 const CUSTOM_ENABLED_KEY = "oem-button-custom-enabled";
 const BUTTON_KEY = "oem-button-values";
